Fix SBT download
- Fixes error "Invalid or Corrupt jar file" when running sbt, by applying equivalent change from https://github.com/databricks/spark-perf/pull/78

Typesafe has deprecated artifactoryonline in favor of bintray, as announced in this blog post.
https://www.typesafe.com/blog/migrating-repos-to-bintray

This results in sbt-launch-lib.bash to fail in downloading SBT, leading to an error:

> Attempting to fetch sbt
> Launching sbt from sbt/sbt-launch-0.13.6.jar
> Error: Invalid or corrupt jarfile sbt/sbt-launch-0.13.6.jar

This change updates the repo URL and fixes the curl command to handle 3xx redirects properly.

Props to @feynmanliang for the pointer to this fix :)